itech_api-main/api/api/requirements.txt
uvicorn
fastapi
pymysql
mysqlclient
sqlalchemy==1.4.32
ldap3
pyjwt
pyyaml
requests
jinja2
python-multipart
